deployment/examples/api-model_in_service/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
colorama==0.4.6 ; platform_system == 'Windows'
coloredlogs==15.0.1
fastapi==0.115.6
flatbuffers==24.12.23
h11==0.14.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.1
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.2
pydantic==2.10.4
pydantic-core==2.27.2
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
typing-extensions==4.12.2
uvicorn==0.34.0
